remove duplicated test case
